refactor: remove plugins/data
