+++ packages/plugins/src/index.ts
@@ -1,2 +1,1 @@
 export * from './plugins';
-export * from './data';
